debian/gnome-terminal.wrapper: Always wait for terminal to exit
GNOME Terminal now uses a dbus-activated service, so calling it would exit
immediately once it dbus-activates the server, while the wrapper is
supposed to mimic xterm behavior and so to wait until the terminal is
explicitly closed.

This is particularly relevant for the XSession file that gdm ships, as in
case of failures it calls x-terminal-emulator (which very likely points to
this script) to open a terminal for debugging purposes, that should stay
visible until the user closes it. But right now we just exit immediately
and the x session is closed, sending the user back to gdm screen.

As per this, always use the --wait (implemented in commit 42f9f797) argument
to call gnome-terminal, so that the behavior is preserved.